updated migration to float

+++ db/migrate/20150417235711_create_courses_points.rb
@@ -1,11 +1,11 @@
 class CreateCoursesPoints < ActiveRecord::Migration
   def change
     create_table :courses_points do |t|
-      t.integer :longitude
-      t.integer :latitude
+      t.float :longitude
+      t.float :latitude
       t.integer :course_id
-      t.integer :altitude
-      t.integer :speed
+      t.float :altitude
+      t.float :speed
 
       t.timestamps null: false
     end
